Refactor compilation like behaviour
A number of different tasks will want errors to be linked to code,
for `next-error` to work, etc. This is handled by `compilation-mode` in
Emacs, so this functionality is called "compile like" behaviour in
`mill-mode`.

Add a utility to run compile like
commands (`mill-mode-run-compile-like`).

Build the user visible compile command on top of this primitive.

Fix up buffer handling so the Mill buffer is created if it doesn't
already exist.

Refactor code into separate files so it's easier to navigate.